fehlende felder hinzugefügt

+++ product/product.go
@@ -1,6 +1,6 @@
 package product
 
 type Product struct {
-	Name  string `json:"name"`
-	Price int    `json:"price"`
+	Name  string  `json:"name"`
+	Price float32 `json:"price"`
 }
